graphslam-engine: Add app-related methods in TGraphSlamResources class
Kickstart a new class (TGraphSlamResources) which holds all the
app-related functions. These deal with the following:

- Validation of user input of the selected deciders/optimizers.
- Printing of the available deciders/optimizers as well as their usage-
  Handy for dumping the deciders descriptions in any existing
  application help messages
- (Static) methods for creating decider/optimizer instances given the
  decider/optimizer name in "string" form - handy for initializing the
  correct object on the fly based on the user input

Note: Files are placed in mrpt-graphslam lib so that they can be used by
any application (not only a native MRPT application) that executes
graphSLAM (e.g. a ROS wrapper for graphslam-engine)
